refactor: Move test files to `fake` package and rename test packages
This commit reorganizes the test files by moving them into a dedicated `fake` package within the `domain` module. It also renames the packages where use case tests reside.

-   **domain/src/test/java/com/ab/domain/fake/:**
    -   Created `fake` package to store fake test data and classes.
    -   Moved `FakeTestData`, `FakeResourceProvider`, and `FakeWeatherForecastRepository` into the `fake` package.
- **domain/src/test/java/com/ab/domain/usecases/:**
    - Renamed package from `com.ab.domain` to `com.ab.domain.usecases`
    -   Moved `InsertWeatherForecastResponseToLocalUseCaseFakeTest` and `GetLastSearchedCityUseCaseFakeTest` to `usecases` package.
