+++ domain/src/test/java/com/ab/domain/usecases/GetLastSearchedCityUseCaseFakeTest.kt
@@ -1,7 +1,9 @@
-package com.ab.domain
+package com.ab.domain.usecases
 
 import com.ab.core.utils.base.Resource
-import com.ab.domain.usecases.GetLastSearchedCityUseCase
+import com.ab.domain.fake.FakeResourceProvider
+import com.ab.domain.fake.FakeTestData
+import com.ab.domain.fake.FakeWeatherForecastRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
@@ -20,25 +22,6 @@ class GetLastSearchedCityUseCaseFakeTest {
         fakeRepository = FakeWeatherForecastRepository()
         fakeResourceProvider = FakeResourceProvider()
         useCase = GetLastSearchedCityUseCase(fakeRepository, fakeResourceProvider)
-    }
-
-    @Test
-    fun invoke_cityAndForecastAvailable_returnsSuccess() = runTest {
-        val cityEntity = FakeTestData.buildCityEntity(id = 1, name = "Cairo")
-        fakeRepository.insertCityToLocal(cityEntity)
-
-        val forecast = FakeTestData.buildForecastEntity(cityId = 1, today = true)
-        val temp = FakeTestData.buildTemperatureEntity(forecastId = forecast.id)
-        val condition = FakeTestData.buildConditionEntity(forecastId = forecast.id)
-
-        fakeRepository.insertWeatherForecastToLocal(forecast)
-        fakeRepository.insertTemperatureToLocal(temp)
-        fakeRepository.insertWeatherConditionToLocal(condition)
-
-        val result = useCase()
-
-        assertTrue(result is Resource.Success)
-        assertEquals("Cairo", (result as Resource.Success).data.name)
     }
 
     @Test
